+++ src/main/java/dev/crossvas/jadexic2c/info/CableInfoProvider.java
@@ -51,8 +51,8 @@ public enum CableInfoProvider implements IHelper<BlockEntity> {
 
                 if (averageOut > 0) {
                     PluginHelper.spacerY(iTooltip, 10);
-                    TextHelper.text(iTooltip, Component.translatable("tooltip.item.ic2.eu_reader.cable_flow", Formatters.EU_FORMAT.format((long) averageOut)).withStyle(ChatFormatting.AQUA));
-                    TextHelper.text(iTooltip, Component.translatable("tooltip.item.ic2.eu_reader.packet_flow", Formatters.EU_FORMAT.format((long) averageOutPacket)).withStyle(ChatFormatting.AQUA));
+                    TextHelper.text(iTooltip, Component.translatable("tooltip.item.ic2.eu_reader.cable_flow", Formatters.EU_FORMAT.format(averageOut)).withStyle(ChatFormatting.AQUA));
+                    TextHelper.text(iTooltip, Component.translatable("tooltip.item.ic2.eu_reader.packet_flow", Formatters.EU_FORMAT.format(averageOutPacket)).withStyle(ChatFormatting.AQUA));
                 }
             }
         }
